feat(vscode): Add Unit Test Generation Classes for Workflow Run  (#6580)
* update statusCode from int to httpStatusCode and add using Systems.Net for c# file generation

* refactor csharp classes to be used by both create from run and create blank unit test

* add description to classes when there isnt one and remove errorResponse from template

* Fix unit test

* add unit tests for process unit test definition and update location of tests in util functions

* updated unit tests for util mocks

* updated unit tests for test set up

* updates to tests for additional mocks

* Fix unit tests

* Remove unused code

---------
