+++ apps/vs-code-designer/src/app/commands/workflows/openMonitoringView/openMonitoringViewForLocal.ts
@@ -149,7 +149,12 @@ export default class OpenMonitoringViewForLocal extends OpenMonitoringViewBase {
         break;
       }
       case ExtensionCommand.createUnitTest: {
-        await createUnitTest(this.context as IAzureConnectorsContext, vscode.Uri.file(this.workflowFilePath), message.runId);
+        await createUnitTest(
+          this.context as IAzureConnectorsContext,
+          vscode.Uri.file(this.workflowFilePath),
+          message.runId,
+          message.definition
+        );
         break;
       }
       case ExtensionCommand.saveBlankUnitTest: {
