+++ apps/vs-code-designer/src/app/commands/workflows/unitTest/__test__/saveBlankUnitTest.test.ts
@@ -1,0 +1,120 @@
+// saveBlankUnitTest.test.ts
+import { describe, test, expect, beforeEach, afterEach, vi } from 'vitest';
+import * as vscode from 'vscode';
+import * as fs from 'fs-extra';
+import * as path from 'path';
+
+// Import the function under test and the utility modules
+import { saveBlankUnitTest } from '../saveBlankUnitTest';
+import * as workspaceUtils from '../../../../utils/workspace';
+import * as projectRootUtils from '../../../../utils/verifyIsProject';
+import * as unitTestUtils from '../../../../utils/unitTests';
+import * as azextUtils from '@microsoft/vscode-azext-utils';
+import { ext } from '../../../../../extensionVariables';
+
+vi.mock('../../../../../extensionVariables', () => ({
+  ext: {
+    outputChannel: {
+      appendLog: vi.fn(),
+    },
+  },
+}));
+
+describe('saveBlankUnitTest', () => {
+  const dummyContext: any = { telemetry: {} };
+  const dummyNode: vscode.Uri = vscode.Uri.file('/dummy/path/to/workflow.json');
+  const dummyUnitTestDefinition = {
+    operationInfo: { some: 'info' },
+    outputParameters: { param: 'value' },
+  };
+  const dummyWorkspaceFolder: vscode.WorkspaceFolder = {
+    uri: vscode.Uri.file('/fake/workspace'),
+    name: 'fakeWorkspace',
+    index: 0,
+  };
+  const dummyProjectPath = '/dummy/project';
+  const dummyWorkflowNodeUri = vscode.Uri.file('/dummy/workflow/node.json');
+  const dummyUnitTestName = 'MyUnitTest';
+  const dummyWorkflowName = 'Workflow1';
+  const dummyUnitTestFolderPath = '/dummy/project/tests/Workflow1/MyUnitTest';
+
+  const dummyPaths = {
+    unitTestFolderPath: dummyUnitTestFolderPath,
+    logicAppName: 'LogicApp1',
+    logicAppFolderPath: '/fake/project/myLogicApp',
+    workflowFolderPath: path.join(dummyProjectPath, 'workflows', dummyWorkflowName),
+    testsDirectory: path.join(dummyProjectPath, 'tests'),
+  };
+
+  beforeEach(() => {
+    // Stub utility functions used in saveBlankUnitTest
+    vi.spyOn(workspaceUtils, 'getWorkspaceFolder').mockResolvedValue(dummyWorkspaceFolder);
+    vi.spyOn(projectRootUtils, 'tryGetLogicAppProjectRoot').mockResolvedValue(dummyProjectPath);
+    vi.spyOn(unitTestUtils, 'parseUnitTestOutputs').mockResolvedValue({} as any);
+    vi.spyOn(unitTestUtils, 'selectWorkflowNode').mockResolvedValue(dummyWorkflowNodeUri);
+    vi.spyOn(unitTestUtils, 'promptForUnitTestName').mockResolvedValue(dummyUnitTestName);
+    vi.spyOn(unitTestUtils, 'getUnitTestPaths').mockReturnValue(dummyPaths);
+    vi.spyOn(unitTestUtils, 'processUnitTestDefinition').mockResolvedValue();
+
+    // Stub directory creation
+    vi.spyOn(fs, 'ensureDir').mockResolvedValue();
+
+    // Stub telemetry logging functions
+    vi.spyOn(unitTestUtils, 'logTelemetry').mockImplementation(() => {});
+    vi.spyOn(unitTestUtils, 'handleError').mockImplementation((context, error, source) => {
+      throw error;
+    });
+    vi.spyOn(unitTestUtils, 'logError').mockImplementation(() => {});
+    vi.spyOn(unitTestUtils, 'logSuccess').mockImplementation(() => {});
+
+    // Stub isMultiRootWorkspace to simulate a valid multi-root environment
+    vi.spyOn(workspaceUtils, 'isMultiRootWorkspace').mockReturnValue(true);
+
+    // Stub the callWithTelemetryAndErrorHandling wrapper used inside saveBlankUnitTest
+    vi.spyOn(azextUtils, 'callWithTelemetryAndErrorHandling').mockImplementation(async (eventName, callback) => {
+      // Directly call the callback passed in to simulate success
+      await callback(dummyContext);
+    });
+
+    // Stub methods used within generateBlankCodefulUnitTest
+    vi.spyOn(unitTestUtils, 'createCsFile').mockResolvedValue();
+    vi.spyOn(unitTestUtils, 'ensureCsprojAndNugetFiles').mockResolvedValue();
+    vi.spyOn(workspaceUtils, 'ensureDirectoryInWorkspace').mockResolvedValue();
+    vi.spyOn(ext.outputChannel, 'appendLog').mockImplementation(() => {});
+  });
+
+  afterEach(() => {
+    vi.restoreAllMocks();
+  });
+
+  test('should successfully create a blank unit test', async () => {
+    await saveBlankUnitTest(dummyContext, dummyNode, dummyUnitTestDefinition);
+
+    // Verify that telemetry was logged indicating a successful process
+    expect(unitTestUtils.logTelemetry).toHaveBeenCalledWith(dummyContext, expect.objectContaining({ unitTestSaveStatus: 'Success' }));
+    // Verify that the unit test name was prompted
+    expect(unitTestUtils.promptForUnitTestName).toHaveBeenCalledTimes(1);
+    // Verify that required directories were ensured to exist
+    expect(fs.ensureDir).toHaveBeenCalled();
+    // Verify that the backend process was invoked via callWithTelemetryAndErrorHandling
+    expect(azextUtils.callWithTelemetryAndErrorHandling).toHaveBeenCalled();
+  });
+
+  test('should throw an error if the workspace is not multi-root', async () => {
+    vi.spyOn(workspaceUtils, 'isMultiRootWorkspace').mockReturnValue(false);
+
+    await expect(saveBlankUnitTest(dummyContext, dummyNode, dummyUnitTestDefinition)).rejects.toThrow(
+      /A multi-root workspace must be open to create unit tests/
+    );
+  });
+
+  test('should log an error and call handleError when an exception occurs', async () => {
+    const testError = new Error('Test error');
+    vi.spyOn(unitTestUtils, 'parseUnitTestOutputs').mockRejectedValueOnce(testError);
+
+    await expect(saveBlankUnitTest(dummyContext, dummyNode, dummyUnitTestDefinition)).rejects.toThrow('Test error');
+
+    // Verify that the error logging function was called
+    expect(unitTestUtils.handleError).toHaveBeenCalled();
+  });
+});
