+++ apps/vs-code-designer/test-setup.ts
@@ -23,6 +23,13 @@ vi.mock('@microsoft/vscode-azext-utils', () => {
     }),
     nonNullProp: vi.fn(),
     nonNullValue: vi.fn(),
+    callWithTelemetryAndErrorHandling: (_key: string, callback: Function) => {
+      // Simply invoke the callback with a fake telemetry context.
+      return callback({ telemetry: { properties: {} } });
+    },
+    parseError: vi.fn(() => {
+      return { message: 'error' };
+    }),
   };
 });
 
@@ -33,12 +40,26 @@ vi.mock('fs', () => ({
   createWriteStream: vi.fn(),
 }));
 
+vi.mock('fs-extra', () => ({
+  writeFile: vi.fn(() => Promise.resolve()),
+  ensureDir: vi.fn(() => Promise.resolve()),
+}));
+
 vi.mock('axios');
 
 vi.mock('vscode', () => ({
-  window: {},
+  window: {
+    showInformationMessage: vi.fn(),
+    showErrorMessage: vi.fn(),
+  },
   workspace: {
     workspaceFolders: [],
+  },
+  Uri: {
+    file: (p: string) => ({ fsPath: p, toString: () => p }),
+  },
+  commands: {
+    executeCommand: vi.fn(),
   },
   EventEmitter: vi.fn().mockImplementation(() => ({
     getUser: vi.fn(),
